backend/requirements.txt
Flask==3.0.0

flask-cors==4.0.0

python-dotenv==1.0.0

requests==2.31.0

gunicorn==23.0.0

crewai==0.165.1

crewai_tools==0.35.0

openai==1.35.13